job hooks wip
